+++ app/src/main/java/com/marekpdev/shoppingapp/ui/favourite/FavouriteViewModel.kt
@@ -1,7 +1,6 @@
 package com.marekpdev.shoppingapp.ui.favourite
 
-import android.util.Log
-import com.marekpdev.shoppingapp.mvi.*
+import com.marekpdev.shoppingapp.ui.base.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
