+++ app/src/main/java/com/marekpdev/shoppingapp/ui/search/sort/SortBottomSheetViewModel.kt
@@ -1,6 +1,6 @@
 package com.marekpdev.shoppingapp.ui.search.sort
 
-import com.marekpdev.shoppingapp.mvi.*
+import com.marekpdev.shoppingapp.ui.base.BaseViewModel
 import com.marekpdev.shoppingapp.ui.search.SearchAction
 import com.marekpdev.shoppingapp.ui.search.SearchCommand
 import com.marekpdev.shoppingapp.ui.search.SearchState
